Tweaked translation
Concatenating parts of the sentence is not a good idea. Concatenation
rules may be language specific.

Also, use the default cancel label for the dialog header. It may look
strange when one label gets translated while the rest of the interface
isn't, but let's hope that it's going to motivate people to provide
translations :)